+++ tcex/tcex_app_decorators.py
@@ -1,5 +1,73 @@
 # -*- coding: utf-8 -*-
 """App Decorators"""
+import datetime
+# import time
+
+
+class Benchmark(object):
+    """Benchmark a function"""
+
+    def __call__(self, fn):
+        """Implement __call__ function for decorator
+
+        Args:
+            fn (function): The decorated function.
+
+        Returns:
+            function: The custom decorator function.
+        """
+
+        def benchmark(app, *args, **kwargs):
+            """Iterate over data, calling the decorated function for each value.
+
+            Args:
+                app (class): The instance of the App class "self".
+            """
+
+            # before = float('{0:.4f}'.format(time.clock()))
+            before = datetime.datetime.now()
+            data = fn(app, *args, **kwargs)
+            # after = float('{0:.4f}'.format(time.clock()))
+            after = datetime.datetime.now()
+            app.tcex.log.debug(
+                'function: "{}", benchmark_time: "{}"'.format(
+                    self.__class__.__name__, after - before
+                )
+            )
+            return data
+
+        return benchmark
+
+
+class Debug(object):
+    """Debug a function"""
+
+    def __call__(self, fn):
+        """Implement __call__ function for decorator
+
+        Args:
+            fn (function): The decorated function.
+
+        Returns:
+            function: The custom decorator function.
+        """
+
+        def debug(app, *args, **kwargs):
+            """Iterate over data, calling the decorated function for each value.
+
+            Args:
+                app (class): The instance of the App class "self".
+            """
+
+            data = fn(app, *args, **kwargs)
+            app.tcex.log.debug(
+                'function: "{}", args: "{}", kwargs: "{}"'.format(
+                    self.__class__.__name__, vars(args), kwargs
+                )
+            )
+            return data
+
+        return debug
 
 
 class InterateOnArg(object):
@@ -26,7 +94,7 @@ class InterateOnArg(object):
             function: The custom decorator function.
         """
 
-        def loop(app, *args):
+        def loop(app, *args, **kwargs):
             """Iterate over data, calling the decorated function for each value.
 
             Args:
@@ -34,6 +102,7 @@ class InterateOnArg(object):
             """
 
             # retrieve data from Redis if variable and always return and array.
+            r = []
             for s in app.tcex.playbook.read(getattr(app.args, self.arg), True):
                 # update the first item in the tuple
                 args_list = list(args)
@@ -42,7 +111,8 @@ class InterateOnArg(object):
                 except IndexError:
                     args_list.append(s)
                 args = tuple(args_list)
-                return fn(app, *args)
+                r.append(fn(app, *args, **kwargs))
+            return r
 
         return loop
 
@@ -69,7 +139,7 @@ class OnException(object):
             function: The custom decorator function.
         """
 
-        def exception(app, *args):
+        def exception(app, *args, **kwargs):
             """Call the function and handle any exception.
 
             Args:
@@ -77,7 +147,7 @@ class OnException(object):
             """
 
             try:
-                return fn(app, *args)
+                return fn(app, *args, **kwargs)
             except Exception as e:
                 app.tcex.log.error('method failure ({})'.format(e))
                 app.tcex.exit(1, self.msg)
@@ -107,7 +177,7 @@ class OnSuccess(object):
             function: The custom decorator function.
         """
 
-        def completion(app, *args):
+        def completion(app, *args, **kwargs):
             """Call the function and handle any exception.
 
             Args:
@@ -115,7 +185,7 @@ class OnSuccess(object):
             """
 
             app.msg = self.msg
-            return fn(app, *args)
+            return fn(app, *args, **kwargs)
 
         return completion
 
@@ -123,15 +193,14 @@ class OnSuccess(object):
 class Output(object):
     """Set exit message on success"""
 
-    def __init__(self, output, append=False):
+    def __init__(self, attribute):
         """Initialize Class Properties
 
         Args:
-            output (str): The name of the output variable.
+            attribute (str): The name of the App attribute to write data.
         """
 
-        self.append = append
-        self.output = output
+        self.attribute = attribute
 
     def __call__(self, fn):
         """Implement __call__ function for decorator
@@ -143,18 +212,21 @@ class Output(object):
             function: The custom decorator function.
         """
 
-        def output(app, *args):
+        def output(app, *args, **kwargs):
             """Call the function and store or append return value.
 
             Args:
                 app (class): The instance of the App class "self".
             """
 
-            data = fn(app, *args)
-            if self.append:
-                getattr(app, self.output).append(data)
+            data = fn(app, *args, **kwargs)
+            attr = getattr(app, self.attribute)
+            if isinstance(data, list) and isinstance(attr, list):
+                getattr(app, self.attribute).extend(data)
+            elif isinstance(attr, list):
+                getattr(app, self.attribute).append(data)
             else:
-                setattr(app, self.output, data)
+                setattr(app, self.attribute, data)
             return data
 
         return output
@@ -185,7 +257,7 @@ class ReadArg(object):
             function: The custom decorator function.
         """
 
-        def read(app, *args):
+        def read(app, *args, **kwargs):
             """Retrieve/Read data from Redis.
 
             Args:
@@ -197,11 +269,9 @@ class ReadArg(object):
             try:
                 args_list[0] = app.tcex.playbook.read(getattr(app.args, self.arg), self.array)
             except IndexError:
-                args_list.append(
-                    app.tcex.playbook.read(getattr(app.args, self.arg), self.array)
-                )
+                args_list.append(app.tcex.playbook.read(getattr(app.args, self.arg), self.array))
             args = tuple(args_list)
-            return fn(app, *args)
+            return fn(app, *args, **kwargs)
 
         return read
 
@@ -209,20 +279,23 @@ class ReadArg(object):
 class WriteOutput(object):
     """Set exit message on success"""
 
-    def __init__(self, var_name, var_type, value=None):
+    def __init__(self, key, variable_type, value=None, overwrite=True):
         """Initialize Class Properties
 
         Args:
-            var_name (str): The name of the playbook output variable.
-            var_type (str): The type for the playbook output variable.  Supported types are:
+            key (str): The name of the playbook output variable.
+            variable_type (str): The type for the playbook output variable.  Supported types are:
                 String, Binary, KeyValue, TCEntity, TCEnhancedEntity, StringArray,
                 BinaryArray, KeyValueArray, TCEntityArray, TCEnhancedEntityArray.
+            value (str): When a value is provided that value will be written instead of the data
+                returned from the function call.
+            overwrite (bool): When True and more than one value is provided for the same variable
+                the previous value will be overwritten.
         """
 
-        self.app_output = 'tc_app_output'  # defined in PlaybookApp class
-        self.var_name = var_name
-        self.var_type = var_type
+        self.key = key
         self.value = value
+        self.variable_type = variable_type
 
     def __call__(self, fn):
         """Implement __call__ function for decorator
@@ -234,18 +307,24 @@ class WriteOutput(object):
             function: The custom decorator function.
         """
 
-        def output(app, *args):
+        def output(app, *args, **kwargs):
             """Call the function and store or append return value.
 
             Args:
                 app (class): The instance of the App class "self".
             """
 
-            data = fn(app, *args)
+            data = fn(app, *args, **kwargs)
+            index = '{}-{}'.format(self.key, self.variable_type)
             if self.value is not None:
-                app.tcex.playbook.add_output(self.var_name, self.value, self.var_type)
+                # store user provided data
+                app.tcex.playbook.add_output(self.key, self.value, self.variable_type)
+            elif app.tcex.playbook.output_data.get(index) and not overwrite:
+                # skip data since a previous value has already been written
+                pass
             else:
-                app.tcex.playbook.add_output(self.var_name, data, self.var_type)
+                # store data returned by function call
+                app.tcex.playbook.add_output(self.key, data, self.variable_type)
             return data
 
         return output